+++ src/common/meta/src/key.rs
@@ -84,7 +84,7 @@ pub mod catalog_name;
 pub mod datanode_table;
 /// TODO(weny):removes id.
 #[allow(unused)]
-pub mod flow_task;
+pub mod flow;
 pub mod schema_name;
 pub mod scope;
 pub mod table_info;
@@ -123,8 +123,8 @@ use table_name::{TableNameKey, TableNameManager, TableNameValue};
 
 use self::catalog_name::{CatalogManager, CatalogNameKey, CatalogNameValue};
 use self::datanode_table::RegionInfo;
-use self::flow_task::flow_task_info::FlowTaskInfoValue;
-use self::flow_task::flow_task_name::FlowTaskNameValue;
+use self::flow::flow_task_info::FlowTaskInfoValue;
+use self::flow::flow_task_name::FlowTaskNameValue;
 use self::schema_name::{SchemaManager, SchemaNameKey, SchemaNameValue};
 use self::table_route::{TableRouteManager, TableRouteValue};
 use self::tombstone::TombstoneManager;
